focus on helpers.

+++ qemu/types.h
@@ -2,8 +2,8 @@
 #define TYPES_H
 #include <stdint.h>
 typedef uint8_t flag;
-typedef uint32_t target_ulong;
-typedef int32_t target_long;
+typedef uint64_t target_ulong;
+typedef int64_t target_long;
 
 #ifndef __cplusplus
 typedef uint8_t bool;
